+++ clox/src/value.c
@@ -43,7 +43,7 @@ void
 value_array_write(ValueArray *array, Value value)
 {
   if (array->capacity < array->count + 1) {
-    size_t old_capacity = array->capacity;
+    int old_capacity = array->capacity;
     array->capacity = GROW_CAPACITY(old_capacity);
     array->values = GROW_ARRAY(Value, array->values, old_capacity,
                                array->capacity);
